% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/atracker_data.R
\name{atracker_data}
\alias{atracker_data}
\title{A function to read ATracker csv data files into one data frame and transform data for analysis}
\usage{
atracker_data(path_to_data)
}
\arguments{
\item{path_to_data}{where to find the csv files}
}
\description{
A function to read ATracker csv data files into one data frame and transform data for analysis
}
\examples{
atracker_data("./data")
}
\keyword{data}
\keyword{load}
\keyword{transform}

