% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_atracker_data.R
\name{read_ATracker_data}
\alias{read_ATracker_data}
\title{A function to read ATracker csv data file}
\usage{
read_ATracker_data(path_to_csv)
}
\arguments{
\item{path_to_csv}{where to find the csv file}
}
\description{
A function to read ATracker csv data file
}
\examples{
read_ATracker_data("./data.csv")
}
\keyword{data}
\keyword{read}

